Amelioration a l'algo de match de points entre 2 frames

+++ Prototype/kinectPower/hand_extractor/hand_2d_parameters.h
@@ -22,7 +22,7 @@ class Hand2dParameters {
 
   struct PotentialTip {
     cv::Point position;
-    int lifetime;
+    float lifetime;
   };
 
   typedef std::vector<PotentialTip> PotentialTipVector;
